Guard against nil reference

+++ collectors/mmlsfileset.go
@@ -175,7 +175,9 @@ func mmlsfileset(fs string, ctx context.Context) (string, error) {
 	cmd.Stdout = &out
 	err := cmd.Run()
 	if ctx.Err() == context.DeadlineExceeded {
-		syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
+		if cmd.Process != nil {
+			syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
+		}
 		return "", ctx.Err()
 	} else if err != nil {
 		return "", err
